requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-voyageai>=0.0.1
pydantic>=2.5.3,<3.0.0
pinecone-client>=3.0.0
python-dotenv==1.0.0
voyageai>=0.1.3
pypdf==3.17.1
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
langsmith>=0.0.87
streamlit>=1.32.0

